refactor - move patches to dedicated cc_physpatch
feature - friction-hash warmup improvement to IVP
feature - controller-list UAF & VHash absent-key fatal patches
