+++ src/features/index.hpp
@@ -5,6 +5,7 @@
 #include "tools/cc_signature.h"
 #include "tools/cc_patch.h"
 #include "features/cc_physrecover.h"
+#include "features/cc_physpatch.h"
 #include "features/cc_engine.h"
 #include "features/cc_profile.h"
 
@@ -15,6 +16,7 @@ namespace CrashCapture {
             #if defined(CC_LINUX)
                 Phys::Recover::Init();
                 Phys::Bind::Init();
+                Phys::Patch::Init();
             #endif
             Engine::Init();
             Profile::Init();
